requirements.txt
django~=4.2.2
djangorestframework
pyyaml
django-cors-headers
requests